# Conflicts:
#   gen-val/src/generation/test/NIST.CVP.Generation.CSHAKE.Tests/ParameterValidatorTests.cs

+++ DB/ACVP_DB/acvp/Stored Procedures/VectorSetJsonGet.sql
@@ -1,6 +1,6 @@
 ﻿
 CREATE PROCEDURE [acvp].[VectorSetJsonGet]
-    @VsId INT,
+    @VsId BIGINT,
     @JsonFileType VARCHAR(20)
 AS
