+++ TaskQueueProcessor/src/NIST.CVP.TaskQueueProcessor/TaskModels/ExecutableTask.cs
@@ -3,7 +3,7 @@ namespace NIST.CVP.TaskQueueProcessor.TaskModels
     public abstract class ExecutableTask
     {
         public long DbId { get; set; }
-        public int VsId { get; set; }
+        public long VsId { get; set; }
         
         public string Error { get; set; }
     }
